@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
         <RootNamespace>Flamui</RootNamespace>
@@ -19,16 +19,13 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
-        <PackageReference Include="ppy.SDL2-CS" Version="1.0.82"/>
-        <PackageReference Include="SkiaSharp" Version="2.88.7" />
-        <PackageReference Include="Svg.Skia" Version="1.0.0.13" />
-        <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="7.0.0" />
         <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
+        <PackageReference Include="FreeTypeSharp" Version="3.0.0" />
+        <PackageReference Include="Silk.NET.Input" Version="2.22.0" />
+        <PackageReference Include="Silk.NET.OpenGL" Version="2.22.0" />
+        <PackageReference Include="Silk.NET.Windowing.Glfw" Version="2.22.0" />
+        <PackageReference Include="StbTrueTypeSharp" Version="1.26.12" />
+        <PackageReference Include="Varena" Version="1.2.0" />
     </ItemGroup>
 
     <ItemGroup>
@@ -40,6 +37,12 @@
     <ItemGroup>
         <None Include="../Flamui.SourceGenerators/bin/Release/netstandard2.0/Flamui.SourceGenerators.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false"/>
         <None Include="../README.md" Pack="true" PackagePath="\"/>
+    </ItemGroup>
+
+    <ItemGroup>
+      <EmbeddedResource Include="Drawing\JetBrainsMono-Regular.ttf" />
+      <EmbeddedResource Include="Drawing\Shaders\main_fragment.glsl" />
+      <EmbeddedResource Include="Drawing\Shaders\main_vertex.glsl" />
     </ItemGroup>
 
 </Project>
