@@ -1,64 +1,118 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.Contracts;
+using System.Drawing;
 using System.Numerics;
-using Flamui.Layouting;
+using System.Runtime.InteropServices;
+using System.Text.Json;
+using Flamui.Drawing;
 using Flamui.UiElements;
-using SkiaSharp;
+using Silk.NET.Maths;
+using Varena;
 
 namespace Flamui;
 
 public class RenderContext
 {
-    public Dictionary<int, RenderSection> RenderSections = new();
+    // public Dictionary<int, RenderSection> RenderSections = new();
     public Stack<int> ZIndexes = new();
 
     public RenderContext()
     {
         ZIndexes.Push(0);
+        Reset();
     }
+
+    private Arena _arena;
+    public Dictionary<int, GrowableArenaBuffer<Command>> CommandBuffers = [];
+
+    private static VirtualArenaManager manager = new();
 
     public void Reset()
     {
-        foreach (var (key, value) in RenderSections)
+        if (_arena == null)
         {
-            value.Renderables.Clear();
+            Console.WriteLine("Creating arena");
+            var virtualBuffer = manager.CreateBuffer("PerFrameArena", (UIntPtr)1_000_000);
+            _arena = new Arena(virtualBuffer);
         }
+
+        CommandBuffers.Clear();
+        _arena.Reset();
+
     }
 
-    public void Add(IRenderFragment renderFragment)
+    public void AddRect(Bounds bounds, UiElement uiElement, ColorDefinition color, float radius = 0)
     {
-        if (!RenderSections.TryGetValue(ZIndexes.Peek(), out var renderSection))
+        var cmd = new Command();
+        cmd.UiElement.Set(_arena, uiElement);
+        cmd.Bounds = bounds;
+        cmd.Radius = radius;
+        cmd.Type = CommandType.Rect;
+        cmd.Color = color;
+
+        Add(cmd);
+    }
+
+    public void AddClipRect(Bounds bounds, float radius = 0)
+    {
+        var cmd = new Command();
+        cmd.Bounds = bounds;
+        cmd.Radius = radius;
+        cmd.Type = CommandType.ClipRect;
+
+        Add(cmd);
+    }
+
+    public void AddText(Bounds bounds, string text, ColorDefinition color, Font font)
+    {
+        var cmd = new Command();
+        cmd.Bounds = bounds;
+        cmd.Type = CommandType.Text;
+        cmd.String.Set(_arena, text);
+        cmd.Color = color;
+        cmd.Font.Set(_arena, font);
+
+        Add(cmd);
+    }
+
+    public void AddMatrix(Matrix4X4<float> matrix)
+    {
+        var cmd = new Command();
+        cmd.Matrix = matrix;
+        cmd.Type = CommandType.Matrix;
+
+        Add(cmd);
+    }
+
+
+    public void Add(Command command)
+    {
+        if (!CommandBuffers.TryGetValue(ZIndexes.Peek(), out var commandBuffer))
         {
-            renderSection = new RenderSection();
-            RenderSections.Add(ZIndexes.Peek(), renderSection);
+            commandBuffer = new GrowableArenaBuffer<Command>(_arena.VirtualBuffer, 20);
+            CommandBuffers.Add(ZIndexes.Peek(), commandBuffer);
         }
 
-        renderSection.Renderables.Add(renderFragment);
+        commandBuffer.Add(command);
     }
 
     public bool RequiresRerender(RenderContext lastRenderContext)
     {
-        foreach (var (key, value) in RenderSections)
+        //maybe go first through everything and check if the sizes match up, and only then go ahead and compare the actual contents
+        foreach (var (key, currentRenderSection) in CommandBuffers)
         {
-            if (!lastRenderContext.RenderSections.TryGetValue(key, out var lastRenderSection))
+            if (!lastRenderContext.CommandBuffers.TryGetValue(key, out var lastRenderSection))
             {
                 return true;
             }
 
-            if (lastRenderSection.Renderables.Count != value.Renderables.Count)
-            {
-                return true;
-            }
+            // if (lastRenderSection.Count != currentRenderSection.Count)
+            // {
+            //     return true;
+            // }
 
-            for (var i = 0; i < value.Renderables.Count; i++)
-            {
-                var renderable = value.Renderables[i];
-                var lastRerenderable = lastRenderSection.Renderables[i];
-
-                if (!renderable.UiEquals(lastRerenderable))
-                {
-                    return true;
-                }
-            }
+            //memcmp...
+            // if (!currentRenderSection.MemCompare(currentRenderSection))
+            //     return false;
         }
 
         return false;
@@ -66,14 +120,55 @@ public class RenderContext
 
     private static int rerenderCount;
 
-    public void Rerender(SKCanvas canvas)
+    private bool isFirstRender = true;
+
+    public void Rerender(Renderer renderer)
     {
-        var sections = RenderSections.OrderBy(x => x.Key).ToList();
+        var canvas = new GlCanvas(renderer);
+
+        canvas.Start();
+
+        var sections = CommandBuffers.OrderBy(x => x.Key).ToList();
 
         foreach (var (_, value) in sections)
         {
-            value.Render(canvas);
+            foreach (var command in value)
+            {
+                if(isFirstRender)
+                    Console.WriteLine(command.Type);
+
+                switch (command.Type)
+                {
+                    case CommandType.Rect:
+                        canvas.Paint.Color = Color.FromArgb(command.Color.Alpha, command.Color.Red, command.Color.Green, command.Color.Blue);
+                        if(command.Radius == 0)
+                            canvas.DrawRect(command.Bounds.X, command.Bounds.Y, command.Bounds.W, command.Bounds.H);
+                        else
+                            canvas.DrawRoundedRect(command.Bounds.X, command.Bounds.Y, command.Bounds.W, command.Bounds.H, command.Radius);
+                        break;
+                    case CommandType.ClipRect:
+                        if(command.Radius == 0)
+                            canvas.ClipRect(command.Bounds.X, command.Bounds.Y, command.Bounds.W, command.Bounds.H);
+                        else
+                            canvas.ClipRoundedRect(command.Bounds.X, command.Bounds.Y, command.Bounds.W, command.Bounds.H, command.Radius);
+                        break;
+                    case CommandType.Text:
+                        canvas.Paint.Font = command.Font.Get();
+                        canvas.Paint.Color = command.Color.ToColor();
+                        canvas.DrawText(command.String.Get(), command.Bounds.X, command.Bounds.Y);
+                        break;
+                    case CommandType.Matrix:
+                        canvas.SetMatrix(command.Matrix);
+                        break;
+                    default:
+                        throw new ArgumentOutOfRangeException(command.Type.ToString());
+                }
+            }
         }
+
+        isFirstRender = false;
+
+        canvas.Flush();
     }
 
     public void SetIndex(int idx)
@@ -87,61 +182,31 @@ public class RenderContext
     }
 }
 
-public class RenderSection
-{
-    public List<IRenderFragment> Renderables = new(1000);
+// public struct Save : IRenderFragment
+// {
+//     public void Render(SKCanvas canvas)
+//     {
+//         canvas.Save();
+//     }
+//
+//     public bool UiEquals(IRenderFragment renderFragment)
+//     {
+//         return renderFragment is Save;
+//     }
+// }
 
-    public void Render(SKCanvas canvas)
-    {
-        foreach (var renderable in Renderables)
-        {
-            renderable.Render(canvas);
-        }
-    }
-}
-
-public interface IClickableFragment
-{
-    public UiElement UiElement { get; set; }
-    public Bounds Bounds { get; set; }
-}
-
-public interface IRenderFragment
-{
-    public void Render(SKCanvas canvas);
-    public bool UiEquals(IRenderFragment renderFragment);
-}
-
-public interface IMatrixable
-{
-    SKPoint ProjectPoint(SKPoint poin);
-}
-
-public struct Save : IRenderFragment
-{
-    public void Render(SKCanvas canvas)
-    {
-        canvas.Save();
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        return renderFragment is Save;
-    }
-}
-
-public struct Restore : IRenderFragment
-{
-    public void Render(SKCanvas canvas)
-    {
-        canvas.Restore();
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        return renderFragment is Restore;
-    }
-}
+// public struct Restore : IRenderFragment
+// {
+//     public void Render(SKCanvas canvas)
+//     {
+//         canvas.Restore();
+//     }
+//
+//     public bool UiEquals(IRenderFragment renderFragment)
+//     {
+//         return renderFragment is Restore;
+//     }
+// }
 
 public struct Bounds
 {
@@ -152,6 +217,7 @@ public struct Bounds
 
     public float Right => X + W;
 
+    [Pure]
     public bool ContainsPoint(Vector2 point)
     {
         var withinX = point.X >= X && point.X <= X + W;
@@ -159,19 +225,11 @@ public struct Bounds
 
         return withinX && withinY;
     }
-
-    public SKRect ToRect()
-    {
-        return SKRect.Create(X, Y, W, H);
-    }
-
-    public bool BoundsEquals(Bounds otherBounds)
-    {
-        return otherBounds.X == X
-            && otherBounds.Y == Y
-            && otherBounds.W == W
-            && otherBounds.H == H;
-    }
+    //
+    // public SKRect ToRect()
+    // {
+    //     return SKRect.Create(X, Y, W, H);
+    // }
 
     public Bounds Inflate(float x, float y)
     {
@@ -190,297 +248,227 @@ public struct Bounds
     }
 }
 
-public struct RectClip : IRenderFragment
+public enum CommandType : byte
 {
-    private static readonly SKRoundRect RoundRect = new();
+    Rect,
+    ClipRect,
+    Text,
+    Matrix
+}
 
-    public required Bounds Bounds { get; set; }
-    public required float Radius;
-    public required SKClipOperation ClipOperation;
+public struct Command
+{
+    public CommandType Type;
+    public ManagedRef<UiElement> UiElement;
+    public ManagedRef<string> String;
+    public Bounds Bounds;
+    public float Radius;
+    public ManagedRef<Font> Font;
+    public ColorDefinition Color;
+    public Matrix4X4<float> Matrix;
+}
 
-    public void Render(SKCanvas canvas)
+public struct ManagedRef<T> where T : class
+{
+    private GCHandle handle;
+
+    public void Set(Arena arena, T value)
     {
-        if (Radius == 0)
+        handle = arena.AddReference(value);
+    }
+
+    [Pure]
+    public T Get()
+    {
+        if (handle.Target != null)
         {
-            canvas.ClipRect(Bounds.ToRect(), ClipOperation, true);
-        }
-        else
-        {
-            RoundRect.SetRect(Bounds.ToRect(), Radius, Radius);
-            canvas.ClipRoundRect(RoundRect, ClipOperation, antialias: true);
-        }
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not RectClip rectClip)
-            return false;
-
-        return rectClip.Bounds.BoundsEquals(Bounds)
-               && rectClip.Radius == Radius
-               && rectClip.ClipOperation == ClipOperation;
-    }
-}
-
-public struct Rect : IRenderFragment, IClickableFragment
-{
-    public required UiElement UiElement { get; set; }
-    public required Bounds Bounds { get; set; }
-    public required float Radius;
-    public required IRenderPaint RenderPaint;
-
-    public void Render(SKCanvas canvas)
-    {
-        if (Radius == 0)
-        {
-            canvas.DrawRect(Bounds.ToRect(), RenderPaint.GetPaint());
-        }
-        else
-        {
-            canvas.DrawRoundRect(Bounds.ToRect(), Radius, Radius, RenderPaint.GetPaint());
-        }
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Rect rect)
-            return false;
-
-        return rect.Bounds.BoundsEquals(Bounds)
-               && rect.Radius == Radius
-               && rect.RenderPaint.PaintEquals(RenderPaint);
-    }
-}
-
-public struct Bitmap : IRenderFragment
-{
-    public required Bounds Bounds;
-    public required SKBitmap SkBitmap;
-    private static readonly SKPaint Paint = Helpers.GetNewAntialiasedPaint();
-
-    public void Render(SKCanvas canvas)
-    {
-        canvas.DrawBitmap(SkBitmap, Bounds.ToRect(), Paint);
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Bitmap bitmap)
-            return false;
-
-        return bitmap.Bounds.BoundsEquals(Bounds);
-    }
-}
-
-public struct Picture : IRenderFragment //ToDo, should also be clickable
-{
-    public required SKPicture SkPicture;
-    public required SKMatrix SkMatrix;
-    public required string Src;
-
-    public void Render(SKCanvas canvas)
-    {
-        canvas.DrawPicture(SkPicture, ref SkMatrix);
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Picture pic)
-            return false;
-
-        return pic.Src == Src && SkMatrix == pic.SkMatrix;
-    }
-}
-
-public struct Text : IRenderFragment //ToDo should also be clickable
-{
-    public required string Content;
-    public required float X;
-    public required float Y;
-    public required IRenderPaint RenderPaint;
-
-    public void Render(SKCanvas canvas)
-    {
-        canvas.DrawText(Content, X, Y, RenderPaint.GetPaint());
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Text text)
-            return false;
-
-        return text.Content == Content && text.X == X && text.Y == Y && text.RenderPaint.PaintEquals(RenderPaint);
-    }
-}
-
-public struct Matrix : IRenderFragment, IMatrixable
-{
-    public required SKMatrix SkMatrix;
-
-    public void Render(SKCanvas canvas)
-    {
-        canvas.SetMatrix(canvas.TotalMatrix.PostConcat(SkMatrix));
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Matrix matrix)
-            return false;
-
-        return matrix.SkMatrix == SkMatrix;
-    }
-
-    public SKPoint ProjectPoint(SKPoint poin)
-    {
-        return SkMatrix.Invert().MapPoint(poin);
-    }
-}
-
-public struct Circle : IRenderFragment //ToDo, should also be clickable
-{
-    public required SKPaint SkPaint;
-    public required SKPoint Pos;
-    public required float Radius;
-
-    public void Render(SKCanvas canvas)
-    {
-        canvas.DrawCircle(Pos, Radius, SkPaint);
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Circle circle)
-            return false;
-
-        return circle.Pos == Pos && circle.Radius == Radius;
-    }
-}
-
-public struct Path : IRenderFragment //ToDo maybe should also be clickable
-{
-    public required SKPoint Start;
-    public required SKPoint StartHandle;
-    public required SKPoint EndHandle;
-    public required SKPoint End;
-    public required SKPaint SkPaint;
-
-    private static SKPath _path = new();
-
-    public void Render(SKCanvas canvas)
-    {
-        _path.MoveTo(Start);
-        _path.CubicTo(StartHandle, EndHandle, End);
-
-        canvas.DrawPath(_path, SkPaint);
-
-        _path.Reset();
-    }
-
-    public bool UiEquals(IRenderFragment renderFragment)
-    {
-        if (renderFragment is not Path path)
-            return false;
-
-        return path.Start == Start && path.StartHandle == StartHandle && path.EndHandle == EndHandle &&
-               path.End == End;
-    }
-}
-
-public interface IRenderPaint
-{
-    SKPaint GetPaint();
-    bool PaintEquals(IRenderPaint renderPaint);
-}
-
-public struct TextPaint : IRenderPaint
-{
-    public required float TextSize;
-    public required SKColor SkColor;
-
-    private static readonly SKPaint Paint = MakeTextPaint();
-
-    public SKPaint GetPaint()
-    {
-        Paint.TextSize = TextSize;
-        Paint.Color = SkColor;
-        return Paint;
-    }
-
-    public bool PaintEquals(IRenderPaint renderPaint)
-    {
-        if (renderPaint is not TextPaint textPaint)
-            return false;
-
-        return textPaint.TextSize == TextSize && textPaint.SkColor == SkColor;
-    }
-
-    private static SKPaint MakeTextPaint()
-    {
-        var paint = new SKPaint();
-        paint.IsAntialias = true;
-        paint.Typeface = SKTypeface.FromFamilyName("Arial", SKFontStyleWeight.Thin, SKFontStyleWidth.Normal,
-            SKFontStyleSlant.Upright);
-        return paint;
-    }
-}
-
-public struct ShadowPaint : IRenderPaint
-{
-    public required SKColor SkColor;
-    public required float ShadowSigma;
-
-    private static readonly Dictionary<float, SKMaskFilter> MaskFilterCache = new();
-    private static readonly SKPaint Paint = Helpers.GetNewAntialiasedPaint();
-
-    public SKPaint GetPaint()
-    {
-        if (!MaskFilterCache.TryGetValue(ShadowSigma, out var maskFilter))
-        {
-            //todo maybe ensure that not no many unused maskfilters get created??? because maskfilters are disposable :) AND immutable grrrr
-            maskFilter = SKMaskFilter.CreateBlur(SKBlurStyle.Outer, ShadowSigma, false);
-            MaskFilterCache.Add(ShadowSigma, maskFilter);
+            return (T)handle.Target;
         }
 
-        Paint.Color = SkColor;
-        Paint.MaskFilter = maskFilter;
-        return Paint;
-    }
-
-    public bool PaintEquals(IRenderPaint renderPaint)
-    {
-        if (renderPaint is not ShadowPaint shadowPaint)
-            return false;
-
-        return shadowPaint.SkColor == SkColor && shadowPaint.ShadowSigma == ShadowSigma;
+        throw new Exception("You sure this should be null????");
     }
 }
 
-public struct PlaintPaint : IRenderPaint
-{
-    public required SKColor SkColor;
-    private static readonly SKPaint Paint = Helpers.GetNewAntialiasedPaint();
+// public struct Bitmap : IRenderFragment
+// {
+//     public required Bounds Bounds;
+//     public required SKBitmap SkBitmap;
+//     private static readonly SKPaint Paint = Helpers.GetNewAntialiasedPaint();
+//
+//     public void Render(SKCanvas canvas)
+//     {
+//         canvas.DrawBitmap(SkBitmap, Bounds.ToRect(), Paint);
+//     }
+//
+//     public bool UiEquals(IRenderFragment renderFragment)
+//     {
+//         if (renderFragment is not Bitmap bitmap)
+//             return false;
+//
+//         return bitmap.Bounds.BoundsEquals(Bounds);
+//     }
+// }
 
-    public SKPaint GetPaint()
-    {
-        Paint.Color = SkColor;
-        return Paint;
-    }
+// public struct Picture : IRenderFragment //ToDo, should also be clickable
+// {
+//     public required SKPicture SkPicture;
+//     public required SKMatrix SkMatrix;
+//     public required string Src;
+//
+//     public void Render(SKCanvas canvas)
+//     {
+//         canvas.DrawPicture(SkPicture, ref SkMatrix);
+//     }
+//
+//     public bool UiEquals(IRenderFragment renderFragment)
+//     {
+//         if (renderFragment is not Picture pic)
+//             return false;
+//
+//         return pic.Src == Src && SkMatrix == pic.SkMatrix;
+//     }
+// }
 
-    public bool PaintEquals(IRenderPaint renderPaint)
-    {
-        if (renderPaint is not PlaintPaint plaintPaint)
-            return false;
 
-        return plaintPaint.SkColor == SkColor;
-    }
-}
+// public struct Circle : IRenderFragment //ToDo, should also be clickable
+// {
+//     public required SKPaint SkPaint;
+//     public required SKPoint Pos;
+//     public required float Radius;
+//
+//     public void Render(SKCanvas canvas)
+//     {
+//         canvas.DrawCircle(Pos, Radius, SkPaint);
+//     }
+//
+//     public bool UiEquals(IRenderFragment renderFragment)
+//     {
+//         if (renderFragment is not Circle circle)
+//             return false;
+//
+//         return circle.Pos == Pos && circle.Radius == Radius;
+//     }
+// }
 
-public static class Helpers
-{
-    public static SKPaint GetNewAntialiasedPaint()
-    {
-        var paint = new SKPaint();
-        paint.IsAntialias = true;
-        return paint;
-    }
-}
+// public struct Path : IRenderFragment //ToDo maybe should also be clickable
+// {
+//     public required SKPoint Start;
+//     public required SKPoint StartHandle;
+//     public required SKPoint EndHandle;
+//     public required SKPoint End;
+//     public required SKPaint SkPaint;
+//
+//     private static SKPath _path = new();
+//
+//     public void Render(SKCanvas canvas)
+//     {
+//         _path.MoveTo(Start);
+//         _path.CubicTo(StartHandle, EndHandle, End);
+//
+//         canvas.DrawPath(_path, SkPaint);
+//
+//         _path.Reset();
+//     }
+//
+//     public bool UiEquals(IRenderFragment renderFragment)
+//     {
+//         if (renderFragment is not Path path)
+//             return false;
+//
+//         return path.Start == Start && path.StartHandle == StartHandle && path.EndHandle == EndHandle &&
+//                path.End == End;
+//     }
+// }
+
+// public struct TextPaint : IRenderPaint
+// {
+//     public required float TextSize;
+//     public required SKColor SkColor;
+//
+//     private static readonly SKPaint Paint = MakeTextPaint();
+//
+//     public SKPaint GetPaint()
+//     {
+//         Paint.TextSize = TextSize;
+//         Paint.Color = SkColor;
+//         return Paint;
+//     }
+//
+//     public bool PaintEquals(IRenderPaint renderPaint)
+//     {
+//         if (renderPaint is not TextPaint textPaint)
+//             return false;
+//
+//         return textPaint.TextSize == TextSize && textPaint.SkColor == SkColor;
+//     }
+//
+//     private static SKPaint MakeTextPaint()
+//     {
+//         var paint = new SKPaint();
+//         paint.IsAntialias = true;
+//         paint.Typeface = SKTypeface.FromFamilyName("Arial", SKFontStyleWeight.Thin, SKFontStyleWidth.Normal,
+//             SKFontStyleSlant.Upright);
+//         return paint;
+//     }
+// }
+
+// public struct ShadowPaint : IRenderPaint
+// {
+//     public required SKColor SkColor;
+//     public required float ShadowSigma;
+//
+//     private static readonly Dictionary<float, SKMaskFilter> MaskFilterCache = new();
+//     private static readonly SKPaint Paint = Helpers.GetNewAntialiasedPaint();
+//
+//     public SKPaint GetPaint()
+//     {
+//         if (!MaskFilterCache.TryGetValue(ShadowSigma, out var maskFilter))
+//         {
+//             //todo maybe ensure that not no many unused maskfilters get created??? because maskfilters are disposable :) AND immutable grrrr
+//             maskFilter = SKMaskFilter.CreateBlur(SKBlurStyle.Outer, ShadowSigma, false);
+//             MaskFilterCache.Add(ShadowSigma, maskFilter);
+//         }
+//
+//         Paint.Color = SkColor;
+//         Paint.MaskFilter = maskFilter;
+//         return Paint;
+//     }
+//
+//     public bool PaintEquals(IRenderPaint renderPaint)
+//     {
+//         if (renderPaint is not ShadowPaint shadowPaint)
+//             return false;
+//
+//         return shadowPaint.SkColor == SkColor && shadowPaint.ShadowSigma == ShadowSigma;
+//     }
+// }
+
+// public struct PlaintPaint : IRenderPaint
+// {
+//     public required SKColor SkColor;
+//     private static readonly SKPaint Paint = Helpers.GetNewAntialiasedPaint();
+//
+//     public SKPaint GetPaint()
+//     {
+//         Paint.Color = SkColor;
+//         return Paint;
+//     }
+//
+//     public bool PaintEquals(IRenderPaint renderPaint)
+//     {
+//         if (renderPaint is not PlaintPaint plaintPaint)
+//             return false;
+//
+//         return plaintPaint.SkColor == SkColor;
+//     }
+// }
+
+// public static class Helpers
+// {
+//     public static SKPaint GetNewAntialiasedPaint()
+//     {
+//         var paint = new SKPaint();
+//         paint.IsAntialias = true;
+//         return paint;
+//     }
+// }
