@@ -1,11 +1,29 @@
-
+using System.Numerics;
 using Flamui.Layouting;
-using SkiaSharp;
+using Silk.NET.Maths;
 
 namespace Flamui.UiElements;
 
 public static class FlexContainerRenderer
 {
+    private static Matrix4X4<float> GetRotationMatrix(FlexContainer flexContainer, Point offset)
+    {
+        Vector2 rotationOffset = flexContainer.Info.RotationPivot switch
+        {
+            RotationPivot.Center => new Vector2(flexContainer.Rect.Width / 2, flexContainer.Rect.Height / 2),
+            RotationPivot.TopLeft => new Vector2(0, 0),
+            RotationPivot.TopRight => new Vector2(flexContainer.Rect.Width, 0),
+            RotationPivot.BottomLeft => new Vector2(0, flexContainer.Rect.Height),
+            RotationPivot.BottomRight => new Vector2(flexContainer.Rect.Width, flexContainer.Rect.Height),
+            _ => throw new ArgumentOutOfRangeException()
+        };
+
+        return Matrix4X4<float>.Identity;
+        //todo
+        // return Matrix4X4 CreateRotationDegrees(flexContainer.Info.Rotation, offset.X + rotationOffset.X,
+        //     offset.Y + rotationOffset.Y);
+    }
+
     public static void Render(RenderContext renderContext, FlexContainer flexContainer, Point offset)
     {
         if (flexContainer.Info.ZIndex != 0)
@@ -13,9 +31,28 @@ public static class FlexContainerRenderer
             renderContext.SetIndex(flexContainer.Info.ZIndex);
         }
 
+        if (flexContainer.Info.Rotation != 0)
+        {
+            renderContext.AddMatrix(GetRotationMatrix(flexContainer, offset));
+        }
+
         if (flexContainer.Info.ClipToIgnore is not null)
         {
-            renderContext.Add(new Restore());
+            // renderContext.Add(new Restore());//todo
+        }
+
+        if (flexContainer.Info.BorderWidth != 0 && flexContainer.Info.BorderColor is {} borderColor)
+        {
+            float borderRadius = flexContainer.Info.Radius + flexContainer.Info.BorderWidth;
+
+            var bounds = new Bounds
+            {
+                X = offset.X - flexContainer.Info.BorderWidth,
+                Y = offset.Y - flexContainer.Info.BorderWidth,
+                W = flexContainer.Rect.Width + 2 * flexContainer.Info.BorderWidth,
+                H = flexContainer.Rect.Height + 2 * flexContainer.Info.BorderWidth,
+            };
+            renderContext.AddRect(bounds, flexContainer, borderColor, borderRadius);
         }
 
         if (flexContainer.Info.Color is { } color)
@@ -26,78 +63,28 @@ public static class FlexContainerRenderer
                 float borderRadius = flexContainer.Info.Radius + flexContainer.Info.BorderWidth;
 
                 //todo replace with readable code or something
-                renderContext.Add(new Rect
+                var bounds = new Bounds
                 {
-                    UiElement = flexContainer,
-                    Bounds = new Bounds
-                    {
-                        X = offset.X - flexContainer.Info.BorderWidth + flexContainer.Info.ShadowOffset.Left,
-                        Y = offset.Y - flexContainer.Info.BorderWidth + flexContainer.Info.ShadowOffset.Top,
-                        H = flexContainer.Rect.Height + 2 * flexContainer.Info.BorderWidth - flexContainer.Info.ShadowOffset.Top - flexContainer.Info.ShadowOffset.Bottom,
-                        W = flexContainer.Rect.Width + 2 * flexContainer.Info.BorderWidth - flexContainer.Info.ShadowOffset.Left - flexContainer.Info.ShadowOffset.Right,
-                    },
-                    Radius = flexContainer.Info.Radius == 0 ? 0 : borderRadius,
-                    RenderPaint = new ShadowPaint
-                    {
-                        ShadowSigma = flexContainer.Info.ShadowSigma,
-                        SkColor = blurColor.ToSkColor()
-                    }
-                });
+                    X = offset.X - flexContainer.Info.BorderWidth + flexContainer.Info.ShadowOffset.Left,
+                    Y = offset.Y - flexContainer.Info.BorderWidth + flexContainer.Info.ShadowOffset.Top,
+                    H = flexContainer.Rect.Height + 2 * flexContainer.Info.BorderWidth -
+                        flexContainer.Info.ShadowOffset.Top - flexContainer.Info.ShadowOffset.Bottom,
+                    W = flexContainer.Rect.Width + 2 * flexContainer.Info.BorderWidth -
+                        flexContainer.Info.ShadowOffset.Left - flexContainer.Info.ShadowOffset.Right,
+                };
+
+                //todo shadowsigma from flexcontainer.Info
+                renderContext.AddRect(bounds, flexContainer, blurColor, flexContainer.Info.Radius == 0 ? 0 : borderRadius);
             }
 
-            renderContext.Add(new Rect
-            {
-                UiElement = flexContainer,
-                Bounds = flexContainer.Rect.ToBounds(offset),
-                Radius = flexContainer.Info.Radius,
-                RenderPaint = new PlaintPaint
-                {
-                    SkColor = color.ToSkColor()
-                }
-            });
-        }
-
-        if (flexContainer.Info.BorderWidth != 0 && flexContainer.Info.BorderColor is {} borderColor)
-        {
-            renderContext.Add(new Save());
-
-            float borderRadius = flexContainer.Info.Radius + flexContainer.Info.BorderWidth;
-
-            renderContext.Add(new RectClip
-            {
-                Bounds = flexContainer.Rect.ToBounds(offset),
-                Radius = flexContainer.Info.Radius,
-                ClipOperation = SKClipOperation.Difference
-            });
-
-            renderContext.Add(new Rect
-            {
-                UiElement = flexContainer,
-                Bounds = new Bounds
-                {
-                    X = offset.X - flexContainer.Info.BorderWidth,
-                    Y = offset.Y - flexContainer.Info.BorderWidth,
-                    W = flexContainer.Rect.Width + 2 * flexContainer.Info.BorderWidth,
-                    H = flexContainer.Rect.Height + 2 * flexContainer.Info.BorderWidth,
-                },
-                Radius = borderRadius,
-                RenderPaint = new PlaintPaint
-                {
-                    SkColor = borderColor.ToSkColor()
-                }
-            });
-
-            renderContext.Add(new Restore());
+            renderContext.AddRect(flexContainer.Rect.ToBounds(offset), flexContainer, color, flexContainer.Info.Radius);
         }
 
         ClipContent(renderContext, flexContainer, flexContainer.Info, offset);
 
         if (flexContainer.Info.ScrollConfigY.CanScroll || flexContainer.Info.ScrollConfigX.CanScroll)
         {
-            renderContext.Add(new Matrix
-            {
-                SkMatrix = SKMatrix.CreateTranslation(-flexContainer.ScrollPosX, -flexContainer.ScrollPosY)
-            });
+            renderContext.AddMatrix(Matrix4X4.CreateTranslation(-flexContainer.ScrollPosX, -flexContainer.ScrollPosY, 0));
         }
 
         foreach (var childElement in flexContainer.Children)
@@ -112,17 +99,21 @@ public static class FlexContainerRenderer
 
         if (flexContainer.Info.ScrollConfigY.CanScroll || flexContainer.Info.ScrollConfigX.CanScroll)
         {
-            renderContext.Add(new Matrix
-            {
-                SkMatrix = SKMatrix.CreateTranslation(flexContainer.ScrollPosX, flexContainer.ScrollPosY)
-            });
+            renderContext.AddMatrix(Matrix4X4.CreateTranslation(flexContainer.ScrollPosX, flexContainer.ScrollPosY, 0));
 
-            flexContainer._scrollBarContainer.UiElement?.Render(renderContext, offset.Add(flexContainer._scrollBarContainer.UiElement.ParentData.Position));
+            flexContainer._scrollBarContainerX.UiElement?.Render(renderContext, offset.Add(flexContainer._scrollBarContainerX.UiElement.ParentData.Position));
+            flexContainer._scrollBarContainerY.UiElement?.Render(renderContext, offset.Add(flexContainer._scrollBarContainerY.UiElement.ParentData.Position));
         }
 
         if (NeedsClip(flexContainer.Info))
         {
-            renderContext.Add(new Restore());
+            // renderContext.Add(new Restore());
+        }
+
+        if (flexContainer.Info.Rotation != 0)
+        {
+            Matrix4X4.Invert(GetRotationMatrix(flexContainer, offset), out var inverse);
+            renderContext.AddMatrix(inverse);
         }
 
         //reapply clip
@@ -137,16 +128,16 @@ public static class FlexContainerRenderer
     private static void ClipContent(RenderContext renderContext, UiElement uiElement, FlexContainerInfo Info, Point offset)
     {
         if (NeedsClip(Info))
-
         {
-            renderContext.Add(new Save());
+            // renderContext.Add(new Save());
 
-            renderContext.Add(new RectClip
-            {
-                Bounds = uiElement.Rect.ToBounds(offset),
-                Radius = Info.Radius,
-                ClipOperation = SKClipOperation.Intersect
-            });
+            renderContext.AddClipRect(uiElement.Rect.ToBounds(offset), Info.Radius);
+            // renderContext.Add(new RectClip
+            // {
+            //     Bounds = ,
+            //     Radius = Info.Radius,
+            //     ClipOperation = SKClipOperation.Intersect
+            // });
         }
     }
 
