@@ -1,26 +1,23 @@
-﻿using System.Collections.Concurrent;
+﻿using Flamui.Drawing;
 using Flamui.Layouting;
 using Flamui.UiElements;
 using Microsoft.Extensions.DependencyInjection;
-using SkiaSharp;
+using Silk.NET.Input;
+using Silk.NET.Windowing;
 
 namespace Flamui;
 
 public partial class UiWindow : IDisposable
 {
-    public readonly IntPtr _windowHandle;
-    private readonly FlamuiComponent _rootComponent;
-    public readonly IServiceProvider ServiceProvider;
-    private readonly IntPtr _openGlContextHandle;
-    private readonly GRContext _grContext;
-    public uint Id;
+    public IWindow Window;
+    private FlamuiComponent _rootComponent;
+    public IServiceProvider ServiceProvider;
     public bool IsDebugWindow;
 
     // private UiContainer? _hoveredContainer;
     private UiElement? _activeContainer;
 
-    public readonly UiElementContainer RootContainer;
-    public readonly ConcurrentQueue<SDL_Event> Events = new();
+    public UiElementContainer RootContainer;
 
     // private UiContainer? HoveredDiv
     // {
@@ -40,8 +37,8 @@ public partial class UiWindow : IDisposable
     //     }
     // }
 
-    private readonly Input _input;
-    private readonly HitTester _hitTester;
+    private Input _input;
+    private HitTester _hitTester;
     private readonly TabIndexManager _tabIndexManager = new();
     public readonly Ui Ui = new();
 
@@ -65,66 +62,36 @@ public partial class UiWindow : IDisposable
 
     public List<UiElement> HoveredElements { get; set; } = new();
     public List<UiElement> OldHoveredElements { get; set; } = new();
-    private readonly RegistrationManager _registrationManager;
+    private RegistrationManager _registrationManager;
 
-    public UiWindow(IntPtr windowHandle, FlamuiComponent rootComponent, IServiceProvider serviceProvider)
+    public UiWindow(IWindow window, FlamuiComponent rootComponent, IServiceProvider serviceProvider)
     {
-        _hitTester = new HitTester(this);
-        _windowHandle = windowHandle;
+        Window = window;
         _rootComponent = rootComponent;
         ServiceProvider = serviceProvider;
-        _registrationManager = ServiceProvider.GetRequiredService<RegistrationManager>();
-        Id = SDL_GetWindowID(_windowHandle);
-        _input = new Input(this);
 
-        _openGlContextHandle = SDL_GL_CreateContext(_windowHandle);
-        if (_openGlContextHandle == IntPtr.Zero)
-        {
-            // Handle context creation error
-            Console.WriteLine($"SDL_GL_CreateContext Error: {SDL_GetError()}");
-            SDL_DestroyWindow(_windowHandle);
-            throw new Exception();
-        }
-
-        var success = SDL_GL_MakeCurrent(_windowHandle, _openGlContextHandle);
-        if (success != 0)
-        {
-            throw new Exception();
-        }
-
-        var glInterface = GRGlInterface.CreateOpenGl(SDL_GL_GetProcAddress);
-
-        _grContext = GRContext.CreateGl(glInterface, new GRContextOptions
-        {
-            AvoidStencilBuffers = true
-        });
-
-        Ui.Window = this;
-
-        RootContainer = new FlexContainer
-        {
-            Id = new UiID("RootElement", "", 0, 0),
-            Window = this
-        };
-
-        DebugPaint = Helpers.GetNewAntialiasedPaint();
-        DebugPaint.Color = C.Blue8.ToSkColor();
-
+        window.Load += OnLoad;
+        window.Update += OnUpdate;
+        window.Render += OnRender;
     }
 
-    private SKPaint DebugPaint;
-
-    public RenderContext LastRenderContext = new();
-    public RenderContext RenderContext = new();
-
-    public void Update()
+    private void OnRender(double obj)
     {
-        var success = SDL_GL_MakeCurrent(_windowHandle, _openGlContextHandle);
-        if (success != 0)
-        {
-            throw new Exception();
-        }
+        Render();
 
+        //ToDo cleanup
+        _input.OnAfterFrame();
+
+        OldHoveredElements.Clear();
+        foreach (var uiContainer in HoveredElements)
+        {
+            OldHoveredElements.Add(uiContainer);
+        }
+        HoveredElements.Clear();
+    }
+
+    private void OnUpdate(double obj)
+    {
         ProcessInputs();
 
         //ToDo cleanup
@@ -136,27 +103,33 @@ public partial class UiWindow : IDisposable
         HitDetection();
 
         BuildUi();
-
-
-        Render();
-
-
-
-        //ToDo cleanup
-        _input.OnAfterFrame();
-
-        OldHoveredElements.Clear();
-        foreach (var uiContainer in HoveredElements)
-        {
-            OldHoveredElements.Add(uiContainer);
-        }
-        HoveredElements.Clear();
-
     }
+
+    private Renderer _renderer = new();
+
+    private void OnLoad()
+    {
+        _hitTester = new HitTester(this);
+        _registrationManager = ServiceProvider.GetRequiredService<RegistrationManager>();
+        _input = new Input(Window);
+
+        Ui.Window = this;
+
+        RootContainer = new FlexContainer
+        {
+            Id = new UiID("RootElement", "", 0, 0),
+            Window = this
+        };
+
+        _renderer.Initialize(Window);
+    }
+
+    public RenderContext LastRenderContext = new();
+    public RenderContext RenderContext = new();
 
     public void Close()
     {
-        SDL_DestroyWindow(_windowHandle);
+
     }
 
     private void HitDetection()
@@ -170,34 +143,26 @@ public partial class UiWindow : IDisposable
         RenderToCanvas();
     }
 
-    private void CreateRenderInstructions()
+    private unsafe void CreateRenderInstructions()
     {
         RootContainer.Render(RenderContext, new Point());
     }
 
     private void RenderToCanvas()
     {
-        SDL_GetWindowSize(_windowHandle, out var width, out var height);
-
-        using var renderTarget = new GRBackendRenderTarget(width, height, 0, 8, new GRGlFramebufferInfo(0, 0x8058));
-        using var surface = SKSurface.Create(_grContext, renderTarget, GRSurfaceOrigin.BottomLeft, SKColorType.Rgba8888);
-
         DrawDebugOverlay(RenderContext);
 
-        var requiresRerender = RenderContext.RequiresRerender(LastRenderContext);
+        // var requiresRerender = RenderContext.RequiresRerender(LastRenderContext);
 
         //todo wtf is happening grrrr it makes 0 sense
-        if (requiresRerender)
+        if (true)
         {
             // Console.WriteLine("rerender");
 
             // var start = Stopwatch.GetTimestamp();
 
-            surface.Canvas.Clear();
 
-            RenderContext.Rerender(surface.Canvas);
-
-            surface.Canvas.Flush();
+            RenderContext.Rerender(_renderer);
 
             // Console.WriteLine(Stopwatch.GetElapsedTime(start).TotalMilliseconds);
 
@@ -207,7 +172,8 @@ public partial class UiWindow : IDisposable
         //swap Render Contexts
         (LastRenderContext, RenderContext) = (RenderContext, LastRenderContext);
 
-        _renderHappened = requiresRerender;
+        _renderHappened = true;
+        // _renderHappened = requiresRerender;
     }
 
     private void DrawDebugOverlay(RenderContext renderContext)
@@ -220,7 +186,7 @@ public partial class UiWindow : IDisposable
             var hoveredElement = HoveredElements.FirstOrDefault(x => x != null);
             if (hoveredElement != null)
             {
-                if (IsMouseButtonPressed(MouseButtonKind.Left))
+                if (IsMouseButtonPressed(MouseButton.Left))
                 {
                     DebugSelectedUiElement = hoveredElement;
                     DebugSelectionModelEnabled = false;
@@ -240,36 +206,36 @@ public partial class UiWindow : IDisposable
 
     private void DebugOutline(RenderContext renderContext, Bounds rect)
     {
-        renderContext.Add(new Save());
-
-        renderContext.Add(new RectClip
-        {
-            Bounds = rect,
-            ClipOperation = SKClipOperation.Difference,
-            Radius = 0
-        });
-
-        renderContext.Add(new Rect
-        {
-            Bounds = rect.Inflate(2, 2),
-            Radius = 0,
-            RenderPaint = new PlaintPaint
-            {
-                SkColor = C.Blue8.ToSkColor()
-            },
-            UiElement = null
-        });
+        // renderContext.Add(new Save());
+        //
+        // renderContext.Add(new RectClip
+        // {
+        //     Bounds = rect,
+        //     ClipOperation = SKClipOperation.Difference,
+        //     Radius = 0
+        // });
+        //
+        // renderContext.Add(new Rect
+        // {
+        //     Bounds = rect.Inflate(2, 2),
+        //     Radius = 0,
+        //     RenderPaint = new PlaintPaint
+        //     {
+        //         SkColor = C.Blue8.ToSkColor()
+        //     },
+        //     UiElement = null
+        // });
     }
 
     private void ProcessInputs()
     {
-        _input.HandleEvents(Events);
+        // _input.HandleEvents(Events);
         _tabIndexManager.HandleTab(this);
     }
 
     private void BuildUi()
     {
-        SDL_GetWindowSize(_windowHandle, out var width, out var height);
+        // SDL_GetWindowSize(Window, out var width, out var height);
 
         Ui.OpenElementStack.Clear();
         Ui.OpenElementStack.Push(RootContainer);
@@ -283,28 +249,28 @@ public partial class UiWindow : IDisposable
         _rootComponent.Build(Ui);
 
         RootContainer.PrepareLayout(Dir.Vertical);
-        RootContainer.Layout(new BoxConstraint(0, width, 0, height));
+        RootContainer.Layout(new BoxConstraint(0, Window.Size.X, 0, Window.Size.Y));
     }
 
     private bool _renderHappened;
 
     public void SwapWindow()
     {
-        var success = SDL_GL_MakeCurrent(_windowHandle, _openGlContextHandle);
-        if (success != 0)
-        {
-            throw new Exception();
-        }
-
-        if(_renderHappened)
-            SDL_GL_SwapWindow(_windowHandle);
-
-        //ToDo
+        // var success = SDL_GL_MakeCurrent(Window, _openGlContextHandle);
+        // if (success != 0)
+        // {
+        //     throw new Exception();
+        // }
+        //
+        // if(_renderHappened)
+        //     SDL_GL_SwapWindow(Window);
+        //
+        // //ToDo
     }
 
     public void Dispose()
     {
-        SDL_GL_DeleteContext(_openGlContextHandle);
-        SDL_DestroyWindow(_windowHandle);
+        // SDL_GL_DeleteContext(_openGlContextHandle);
+        // SDL_DestroyWindow(Window);
     }
 }
